+++ src/main/java/io/vacco/jtinn/error/JtMeanSquaredError.java
@@ -6,11 +6,11 @@ public class JtMeanSquaredError implements JtErrorFn {
 
   private static final long serialVersionUID = JtSchema.version;
 
-  @Override public double of(double a, double b) {
+  @Override public float of(float a, float b) {
     return 0.5f * (a - b) * (a - b);
   }
 
-  @Override public double pd(double act, double target) {
+  @Override public float pd(float act, float target) {
     return act - target;
   }
