+++ src/main/java/io/vacco/jtinn/net/JtNetwork.java
@@ -28,30 +28,30 @@ public class JtNetwork implements Serializable {
     return this;
   }
 
-  private void activate(double[] in, boolean update, JtLayer l) {
-    double z;
-    double[] out = update ? l.a : l.ar;
+  private void activate(float[] in, boolean update, JtLayer l) {
+    float z;
+    float[] out = update ? l.a : l.ar;
 
     if (!update) { System.arraycopy(l.a, 0, l.ar, 0, l.a.length); }
     for (int j = 0; j < l.size(); j++) {
       z = 0;
       for (int a = 0; a < l.weightSize(); a++) {
-        double aj = in[a];
-        double wj = l.w[j][a];
+        float aj = in[a];
+        float wj = l.w[j][a];
         z = z + aj * wj;
       }
       out[j] = l.actFn.apply(z + l.b[j]);
     }
   }
 
-  private void forward(double[] in, boolean update) {
+  private void forward(float[] in, boolean update) {
     activate(in, update, layerSpec[0]);
     for (int i = 1; i < layerSpec.length; i++) {
       activate(update ? layerSpec[i - 1].a : layerSpec[i - 1].ar, update, layerSpec[i]);
     }
   }
 
-  private void bp1(double[] target, JtOutputLayer l) {
+  private void bp1(float[] target, JtOutputLayer l) {
     JtArrays.checkSize(target, l.a);
     for (int j = 0; j < l.size(); j++) {
       l.δ[j] = l.errFn.pd(l.a[j], target[j]) * l.actFn.pd(l.a[j]);
@@ -59,7 +59,7 @@ public class JtNetwork implements Serializable {
   }
 
   private void bp2(JtLayer l, JtLayer lp1) {
-    double d;
+    float d;
     for (int j = 0; j < l.size(); j++) {
       d = 0;
       for (int k = 0; k < lp1.size(); k++) {
@@ -69,7 +69,7 @@ public class JtNetwork implements Serializable {
     }
   }
 
-  private void backProp(double[] in, double[] tg) {
+  private void backProp(float[] in, float[] tg) {
     bp1(tg, getOutput());
     for (int l = layerSpec.length - 2; l >= 0; l--) {
       bp2(layerSpec[l], layerSpec[l + 1]);
@@ -80,23 +80,23 @@ public class JtNetwork implements Serializable {
     updater.apply(in, layerSpec[0]);
   }
 
-  public double totalError(double[] out) {
+  public float totalError(float[] out) {
     JtOutputLayer ol = getOutput();
     JtArrays.checkSize(out, ol.a);
-    double dt = 0;
+    float dt = 0;
     for (int j = 0; j < out.length; j++) {
       dt = dt + ol.errFn.of(ol.a[j], out[j]);
     }
     return dt;
   }
 
-  public double train(double[] in, double[] out) {
+  public float train(float[] in, float[] out) {
     forward(in, true);
     backProp(in, out);
     return totalError(out);
   }
 
-  public double[] estimate(double[] in) {
+  public float[] estimate(float[] in) {
     forward(in, false);
     return getOutput().ar;
   }
