+++ src/main/java/io/vacco/jtinn/net/JtUpdater.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 
 public interface JtUpdater extends Serializable {
 
-  void apply(double[] lm1a, JtLayer l);
+  void apply(float[] lm1a, JtLayer l);
   void onEpochEnd(int epoch);
 
 }
